Fixed call of method getCapabilities()
