+++ src/Access/CapabilitiesAccess.php
@@ -30,6 +30,22 @@ class CapabilitiesAccess implements AccessInterface
      * @var array
      */
     protected $_userActionCapabilities = [];
+    
+    /**
+     * Controller action(s) capabilities
+     *
+     * @var array
+     */
+    protected $_controllerActionCapabilites = [];
+    
+     /**
+     * Non-assigned actions
+     *
+     * @var array
+     */
+    protected $_nonAssignedActions = [
+        'add'
+    ];
 
     /**
      *  CheckAccess Capabilities
@@ -44,7 +60,7 @@ class CapabilitiesAccess implements AccessInterface
 
         $actionCapabilities = [];
         if (!empty($url['action'])) {
-            $actionCapabilities = static::_getCapabilitiesTable()->getCapabilities($controllerName, [$url['action']]);
+            $actionCapabilities = $this->getCapabilities($controllerName, [$url['action']]);
         }
 
         // if action capabilities is empty, means that current controller or action are skipped
@@ -65,6 +81,46 @@ class CapabilitiesAccess implements AccessInterface
         }
 
         return false;
+    }
+    
+    /**
+     * Returns Controller permission capabilities.
+     *
+     * @param  string $controllerName Controller name
+     * @param  array  $actions        Controller actions
+     * @return array
+     */
+    public function getCapabilities($controllerName = null, array $actions = [])
+    {
+        $result = [];
+
+        if (is_null($controllerName) || !is_string($controllerName)) {
+            return $result;
+        }
+
+        $skipControllers = [];
+        if (is_callable([$controllerName, 'getSkipControllers'])) {
+            $skipControllers = $controllerName::getSkipControllers();
+        }
+
+        if (in_array($controllerName, $skipControllers)) {
+            return $result;
+        }
+
+        $actions = Utils::getActions($controllerName, $actions);
+
+        if (empty($actions)) {
+            return $result;
+        }
+
+        // get controller table instance
+        $controllerTable = Utils::getControllerTableInstance($controllerName);
+
+        return $this->_getCapabilities(
+            Utils::generateCapabilityControllerName($controllerName),
+            $actions,
+            static::_getCapabilitiesTable()->getTableAssignationFields($controllerTable)
+        );
     }
 
     /**
@@ -157,4 +213,61 @@ class CapabilitiesAccess implements AccessInterface
 
         return static::$_capabilitiesTable;
     }
+
+    /**
+     * Method that generates capabilities for specified controller's actions.
+     * Capabilities included are full or owner access types.
+     *
+     * @param  string $controllerName    Controller name
+     * @param  array  $actions           Controller actions
+     * @param  array  $assignationFields Table assignation fields (example: assigned_to)
+     * @return array
+     */
+    protected function _getCapabilities($controllerName, array $actions, array $assignationFields = [])
+    {
+        $key = implode('.', $actions);
+        if (!empty($this->_controllerActionCapabilites[$controllerName][$key])) {
+            return $this->_controllerActionCapabilites[$controllerName][$key];
+        }
+
+        $result = [];
+        foreach ($actions as $action) {
+            // generate action's full (all) type capabilities
+            $result[Utils::CAP_TYPE_FULL][] = new Cap(
+                Utils::generateCapabilityName($controllerName, $action),
+                [
+                    'label' => Utils::generateCapabilityLabel($controllerName, $action . '_all'),
+                    'description' => Utils::generateCapabilityDescription(
+                        $controllerName,
+                        Utils::humanizeActionName($action)
+                    )
+                ]
+            );
+            // skip rest of the logic if assignment fields are not found
+            // or if current action does not support assignment (Example: add / create)
+            if (empty($assignationFields) || in_array($action, $this->_nonAssignedActions)) {
+                continue;
+            }
+
+            // generate action's owner (assignment field) type capabilities
+            foreach ($assignationFields as $assignationField) {
+                $result[Utils::CAP_TYPE_OWNER][] = new Cap(
+                    Utils::generateCapabilityName($controllerName, $action . '_' . $assignationField),
+                    [
+                        'label' => Utils::generateCapabilityLabel($controllerName, $action . '_' . $assignationField),
+                        'description' => Utils::generateCapabilityDescription(
+                            $controllerName,
+                            Utils::humanizeActionName($action) . ' if owner (' . Inflector::humanize($assignationField) . ')'
+                        ),
+                        'field' => $assignationField
+                    ]
+                );
+            }
+        }
+
+        $this->_controllerActionCapabilites[$controllerName][$key] = $result;
+
+        return $this->_controllerActionCapabilites[$controllerName][$key];
+    }
+
 }
